Update code (#798)
* update network for lightweight models on mobile or CPU

* simplify eval input parameters

* update

* update yolov6lite README.md

* fix bug

* fix resume loss&mAP bug

* update README

---------

+++ yolov6/models/losses/loss_distill_ns.py
@@ -140,7 +140,7 @@ class ComputeLoss:
 
         # cls loss
         target_labels = torch.where(fg_mask > 0, target_labels, torch.full_like(target_labels, self.num_classes))
-        one_hot_label = F.one_hot(target_labels, self.num_classes + 1)[..., :-1]
+        one_hot_label = F.one_hot(target_labels.long(), self.num_classes + 1)[..., :-1]
         loss_cls = self.varifocal_loss(pred_scores, target_scores, one_hot_label)
    
         target_scores_sum = target_scores.sum()
